move netstack package to server directory
